requirements.txt
dlt
google-cloud-storage
pyyaml
requests
pandas
